use IChart/IMatch for run lookups
